Force rerender of player rows
The location of every player in a row changes when a player is
either added or removed from that row. This is constrast to most lists
React is used with, like a todo list where the items above the new or
removed one stay where they are.

Because of this, I need the full row to rerender every time to make sure
it's laid out correctly. Using a random row key achieves this in a janky
way, but I couldn't find a more official one.