Use target position directly
The position of the player targets seem to be relative to the top-left
board point already. Not totally sure why it would've been different
from the room stack; maybe that thing's layer is automatically
offsetting itself to wrap the content? That would explain why its
drag event coordinates are (0,0). I'll play around with it.