Move dice to sidebar (#68)
I created my initial floating dice design based on D&D Beyond. However, since that's the only place the roll results are shown (unlike like in D&D Beyond where there's a notification sent to the party), it's clunky to have it be closable. Putting it in the sidebar makes it clearer and always accessible.