Keep name as query param in dev mode
In production, I use localStorage to keep the player's name around in
case they need to refresh for some reason. I can't do that in
development because then I wouldn't be able to join the same game from
multiple tabs as different people.

So in development, I'll keep the name as a query param so I can refresh
freely.